apps/engine/agents/reasoner.py
from typing import Annotated, TypedDict, List
from langgraph.graph import StateGraph, END
from langchain_openai import ChatOpenAI
from langchain_core.messages import BaseMessage, HumanMessage
import os

class AgentState(TypedDict):
    wallet_address: str
    token_address: Optional[str]
    features: dict
    coin_analysis: str
    wallet_analysis: str
    timing_analysis: str
    final_verdict: str
    trust_score: int
    reasoning_trace: List[str]

class MirageReasoner:
    def __init__(self):
        self.llm = ChatOpenAI(model="gpt-4o", api_key=os.getenv("OPENAI_API_KEY"))
        self.workflow = StateGraph(AgentState)
        
        # Define nodes
        self.workflow.add_node("coin_agent", self.coin_agent)
        self.workflow.add_node("wallet_agent", self.wallet_agent)
        self.workflow.add_node("timing_agent", self.timing_agent)
        self.workflow.add_node("aggregator", self.aggregator)
        
        # Define edges
        self.workflow.set_entry_point("coin_agent")
        self.workflow.add_edge("coin_agent", "wallet_agent")
        self.workflow.add_edge("wallet_agent", "timing_agent")
        self.workflow.add_edge("timing_agent", "aggregator")
        self.workflow.add_edge("aggregator", END)
        
        self.app = self.workflow.compile()

    async def coin_agent(self, state: AgentState):
        # Reasoning about token safety and creator history
        prompt = f"Analyze the following token features: {state['features']}. Focus on creator history and tokenomics."
        response = await self.llm.ainvoke([HumanMessage(content=prompt)])
        return {"coin_analysis": response.content}

    async def wallet_agent(self, state: AgentState):
        # Reasoning about wallet funding and clusters
        prompt = f"Analyze the following wallet features: {state['features']}. Focus on funding sources and cluster behavior."
        response = await self.llm.ainvoke([HumanMessage(content=prompt)])
        return {"wallet_analysis": response.content}

    async def timing_agent(self, state: AgentState):
        # Reasoning about transaction timing
        prompt = f"Analyze the following timing features: {state['features']}. Focus on block timing and entropy."
        response = await self.llm.ainvoke([HumanMessage(content=prompt)])
        return {"timing_analysis": response.content}

    async def aggregator(self, state: AgentState):
        # Final verdict and trust score
        prompt = f"Based on the analysis:\nCoin: {state['coin_analysis']}\nWallet: {state['wallet_analysis']}\nTiming: {state['timing_analysis']}\nGive a final verdict (COPY/AVOID/UNCERTAIN) and a trust score (0-100)."
        response = await self.llm.ainvoke([HumanMessage(content=prompt)])
        # Parsing logic would go here
        return {"final_verdict": "UNCERTAIN", "trust_score": 50, "reasoning_trace": [state['coin_analysis'], state['wallet_analysis'], state['timing_analysis']]}
