apps/engine/requirements.txt
fastapi
uvicorn
langgraph
langchain-openai
pydantic
python-dotenv
httpx
psycopg2-binary
redis
